Combine xdrlib tests
